backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-multipart==0.0.9
SQLAlchemy==2.0.35
psycopg[binary,pool]>=3.2.1
alembic==1.13.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pgvector==0.3.2
orjson==3.10.7
